Scroll cursor into view after pull (#2293)
